Move `hash_evm_bytecode` to utils
